changes keyword to established convention
While it's not uncommon for extensions to use a single word as their keyword, other conventions also exist: domains names, and short-hand handles using the @ character. Within the space of Brave Search extension users, there is greater familiarity with the short-hand convention, so we will be following that from now on. This is technically a major change, which merits bumping the version up to 2.0.0, since it breaks the previous "brave" keyword.